// helpers/schema.tsx
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;
// ...
export type ReviewSentiment = "mixed" | "negative" | "neutral" | "positive";

export type Timestamp = ColumnType<Date, Date | string, Date | string>;
// ...
export interface Reviews {
  businessId: number;
  content: string;
  createdAt: Generated<Timestamp>;
  id: Generated<number>;
  rating: number;
  reviewerName: Generated<string>;
  sentiment: Generated<ReviewSentiment | null>;
}
// ...
export const ReviewSentimentArrayValues: [ReviewSentiment, ...ReviewSentiment[]] = ["mixed","negative","neutral","positive"];